% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{app_header}
\alias{app_header}
\title{App Header UI}
\usage{
app_header()
}
\description{
This function creates the header for the RiskIntroApp, including scripts,
styles, and custom message handlers.
}
\keyword{internal}
