% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-config_status.R
\name{build_config_status}
\alias{build_config_status}
\title{Build a Configuration Status}
\usage{
build_config_status(
  value,
  msg,
  error = NULL,
  warnings = NULL,
  warnings_msg = NULL
)
}
\arguments{
\item{value}{boolean value, FALSE for something wrong
in the configuration. TRUE for success.}

\item{msg}{string value, its purpose is to be displayed
if \code{value} is FALSE or TRUE.}

\item{error}{error object or string to display if value is FALSE and there is an error. Often the
\code{res$error} of \code{\link[=safe_and_quiet]{safe_and_quiet()}}.}

\item{warnings}{character vector of warnings. Often the
\code{res$warnings} of \code{\link[=safe_and_quiet]{safe_and_quiet()}}.}

\item{warnings_msg}{Alternative msg value for when value is TRUE and warnings are present.}
}
\description{
Build a configuration status from a boolean and a message to display alongside
it.
}
\seealso{
\code{\link[=report_config_status]{report_config_status()}}
}
\keyword{internal}
