% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-configIsValid.R
\name{config_is_valid}
\alias{config_is_valid}
\title{Validate Configuration}
\usage{
config_is_valid(x, ...)
}
\arguments{
\item{x}{A string with the method to call}

\item{...}{parameters to validate}
}
\value{
A config_status object (logical with message attribute)
}
\description{
Generic function to validate configuration based on validator type.
}
\examples{
config_is_valid.test <- function(x, ...){
params <- list(...)
if(is.null(params$a)) {
  return(FALSE)
}
if(is.null(params$b)) {
  return(FALSE)
}
if(is.null(params$c)) {
  return(FALSE)
}
TRUE
}

config_is_valid("test", a = 1, b = 2, c = 3)

}
