% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-import-epi-units.R
\name{importEpiUnitsUI}
\alias{importEpiUnitsUI}
\title{Epidemiological Units Import Module UI}
\usage{
importEpiUnitsUI(id)
}
\arguments{
\item{id}{Character string. The namespace id for the module.}
}
\value{
An action button for importing epidemiological units.
}
\description{
Creates a simple import button for uploading geospatial epidemiological units data.
}
