% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-config_status.R
\name{is_error}
\alias{is_error}
\title{Checks if an Object is an Error}
\usage{
is_error(x)
}
\arguments{
\item{x}{The object to test. Can be of any type.}
}
\description{
Determines whether an object is an error, specifically if it inherits from
the classes \code{error} or \code{try-error}. For use with resulting values of \code{\link[=safe_and_quite]{safe_and_quite()}}.
}
\keyword{internal}
