% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-onstart.R
\name{onstart}
\alias{onstart}
\title{App onStart function}
\usage{
onstart()
}
\description{
This function is called when the app starts. It sets options for the app and
adds a resource path for the app's static files.
}
