% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-config_status.R
\name{report_config_status}
\alias{report_config_status}
\title{Report a Configuration Status}
\usage{
report_config_status(status, in_panel = TRUE)
}
\arguments{
\item{status}{an object created by \code{build_config_status}}
}
\value{
if \code{status} is not null a shiny::tagList containing an icon and message.
If \code{status} is NULL returns NULL for ease of use in \code{\link[shiny:renderUI]{shiny::renderUI()}}.
}
\description{
Render a simple Shiny-friendly UI element that shows
a green checkmark with the message if valid, or a red cross
with the message if invalid.
}
\keyword{internal}
