% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-configIsValid.R
\name{report_warning}
\alias{report_warning}
\title{Report CLI warning messages in HTML}
\usage{
report_warning(msg)
}
\arguments{
\item{msg}{raw msg string, generally captured by safe_and_quiet()}
}
\value{
HTML object to use with renderUI.
}
\description{
Converts CLI warning messages from riskintroanalysis functions into
pretty UI outputs to use with renderUI and uiOutput.
}
\keyword{internal}
