% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-run.R
\name{run_riskintro}
\alias{run_riskintro}
\title{Run the application locally}
\usage{
run_riskintro(port = 9999, launch.browser = TRUE)
}
\arguments{
\item{port}{used port}

\item{launch.browser}{launch a browser when opening the application}
}
\description{
Run the shiny application locally.
}
