% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-configIsValid.R
\name{safe_and_quiet}
\alias{safe_and_quiet}
\title{Capture all function results and side effects}
\usage{
safe_and_quiet(.fun, ...)
}
\arguments{
\item{.fun}{function to run safely and quietly (caputuring results)}

\item{...}{argments to pass to \code{.fun}}
}
\description{
A combination of \code{\link[purrr:safely]{purrr::safely()}} in case of errors and \link[purrr:quietly]{purrr::quietly}
in case of warnings.
}
\details{
Errors are generally passed to isConfigValid reactives and warnings are passed
to report_warnings
}
\keyword{internal}
